Fix 2026-08-25 bug-report findings F1-F6 across eggserve-core
- F1: validate_components guards its conservative second percent-decode
  with a '%' pre-check, removing N heap allocations per request while
  preserving the literal-%2e%2e denial (behavior pinned by tests)
- F2: serve_connection_with_runtime_state takes Arc<RuntimeConfig>,
  eliminating a deep config clone plus Arc alloc per accepted connection
  (experimental server module; internal callers and tests updated)
- F3: new FileRange::try_new fallible constructor so embedders can avoid
  len()'s documented overflow panic; unit tests cover the boundaries
- F4: Windows handle-relative opens build NtUnicodeString via a shared
  helper that rejects names whose UTF-16 encoding exceeds NT's 16-bit
  length fields with WindowsFsError::NameTooLong instead of silently
  truncating through `as u16`; length/NUL-split documented
- F6: planner test helper make_file_with_size takes usize, dropping the
  truncating u64 cast

F5 (exact-fit NextEntryOffset rejection in the directory-buffer parser)
intentionally unchanged: safe-side strictness per the Win32 contract,
deferred to Windows VM qualification. Recorded in Plan 145.

Verification: fmt/clippy clean, workspace tests 1461 passed (2 ignored),
TLS lint + tests clean (128 passed), x86_64-pc-windows-msvc lib check
clean.

+++ crates/eggserve-core/src/primitives/planner.rs
@@ -580,9 +580,9 @@ mod tests {
     use proptest::prelude::*;
     use std::io::Write;
 
-    fn make_file_with_size(size: u64) -> tempfile::NamedTempFile {
+    fn make_file_with_size(size: usize) -> tempfile::NamedTempFile {
         let mut tmp = tempfile::NamedTempFile::new().unwrap();
-        let data = vec![0u8; size as usize];
+        let data = vec![0u8; size];
         tmp.write_all(&data).unwrap();
         tmp.flush().unwrap();
         tmp
@@ -1583,7 +1583,7 @@ mod tests {
 
     #[test]
     fn property_etag_format() {
-        let sizes = [0u64, 1, 42, 1024, 1024 * 1024];
+        let sizes = [0usize, 1, 42, 1024, 1024 * 1024];
         for size in sizes {
             let tmp = make_file_with_size(size);
             let meta = std::fs::metadata(tmp.path()).unwrap();
@@ -1707,7 +1707,7 @@ mod tests {
         }
 
         #[test]
-        fn generate_etag_never_panics(size in 0u64..=1_000_000) {
+        fn generate_etag_never_panics(size in 0usize..=1_000_000) {
             let tmp = make_file_with_size(size);
             let meta = std::fs::metadata(tmp.path()).unwrap();
             let _ = generate_etag(&meta);
